changed: split VFP classes in base (ebos) and legacy
to get rid of eigen usage in ebos based classes
