+++ opm/autodiff/StandardWells_impl.hpp
@@ -22,8 +22,8 @@
 #include <opm/autodiff/StandardWells.hpp>
 #include <opm/autodiff/WellDensitySegmented.hpp>
 
-#include <opm/autodiff/VFPInjProperties.hpp>
-#include <opm/autodiff/VFPProdProperties.hpp>
+#include <opm/autodiff/VFPInjPropertiesLegacy.hpp>
+#include <opm/autodiff/VFPProdPropertiesLegacy.hpp>
 #include <opm/autodiff/WellHelpers.hpp>
 
 
@@ -96,7 +96,7 @@ namespace Opm
     StandardWells::init(const BlackoilPropsAdFromDeck* fluid_arg,
                         const std::vector<bool>* active_arg,
                         const std::vector<PhasePresence>* pc_arg,
-                        const VFPProperties*  vfp_properties_arg,
+                        const VFPProperties<VFPInjPropertiesLegacy,VFPProdPropertiesLegacy>*  vfp_properties_arg,
                         const double gravity_arg,
                         const Vector& depth_arg)
     {
