+++ opm/autodiff/VFPProdProperties.hpp
@@ -32,9 +32,6 @@
 
 namespace Opm {
 
-template <class Scalar>
-class AutoDiffBlock;
-
 /**
  * Class which linearly interpolates BHP as a function of rate, tubing head pressure,
  * water fraction, gas fraction, and artificial lift for production VFP tables, and similarly
@@ -42,8 +39,6 @@ class AutoDiffBlock;
  */
 class VFPProdProperties {
 public:
-    typedef AutoDiffBlock<double> ADB;
-
     /**
      * Empty constructor
      */
@@ -61,48 +56,8 @@ public:
      * Takes *no* ownership of data.
      * @param prod_tables A map of different VFPPROD tables.
      */
-    explicit VFPProdProperties(const std::map<int, std::shared_ptr<const VFPProdTable> >& prod_tables);
-
-    /**
-     * Linear interpolation of bhp as function of the input parameters.
-     * @param table_id Table number to use
-     * @param wells Wells structure with information about wells in qs
-     * @param qs Flow quantities
-     * @param thp Tubing head pressure
-     * @param alq Artificial lift or other parameter
-     *
-     * @return The bottom hole pressure, interpolated/extrapolated linearly using
-     * the above parameters from the values in the input table.
-     */
-    ADB bhp(const std::vector<int>& table_id,
-            const Wells& wells,
-            const ADB& qs,
-            const ADB& thp,
-            const ADB& alq) const;
-
-    /**
-     * Linear interpolation of bhp as a function of the input parameters given as ADBs
-     * Each entry corresponds typically to one well.
-     * @param table_id Table number to use. A negative entry (e.g., -1)
-     *                 will indicate that no table is used, and the corresponding
-     *                 BHP will be calculated as a constant -1e100.
-     * @param aqua Water phase
-     * @param liquid Oil phase
-     * @param vapour Gas phase
-     * @param thp Tubing head pressure
-     * @param alq Artificial lift or other parameter
-     *
-     * @return The bottom hole pressure, interpolated/extrapolated linearly using
-     * the above parameters from the values in the input table, for each entry in the
-     * input ADB objects.
-     */
-    ADB bhp(const std::vector<int>& table_id,
-            const ADB& aqua,
-            const ADB& liquid,
-            const ADB& vapour,
-            const ADB& thp,
-            const ADB& alq) const;
-
+    using ProdTable = std::map<int, std::shared_ptr<const VFPProdTable> >;
+    explicit VFPProdProperties(const ProdTable& prod_tables);
 
     /**
      * Linear interpolation of bhp as a function of the input parameters given as
@@ -210,7 +165,7 @@ public:
         return m_tables.empty();
     }
 
-private:
+protected:
     // Map which connects the table number with the table itself
     std::map<int, const VFPProdTable*> m_tables;
 };
